add categories list dialog
